Renaming input_parameters to machine_parameters
