--- __doc/source/_modules/interfaces.input_parameters.rst
@@ -1,34 +1,0 @@
-interfaces.input\_parameters package
-========================================================
-
-interfaces.input\_parameters.ring module
-------------------------------------------------------------
-
-.. automodule:: blond_common.interfaces.input_parameters.ring
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    
-interfaces.input\_parameters.ring\_section module
-------------------------------------------------------------
-
-.. automodule:: blond_common.interfaces.input_parameters.ring_section
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-interfaces.input\_parameters.rf\_parameters module
-----------------------------------------------------------------------
-
-.. automodule:: blond_common.interfaces.input_parameters.rf_parameters
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-interfaces.input\_parameters.rf\_parameters\_options module
--------------------------------------------------------------------------------
-
-.. automodule:: blond_common.interfaces.input_parameters.rf_parameters_options
-    :members:
-    :undoc-members:
-    :show-inheritance:
