+++ __doc/source/_modules/interfaces.machine_parameters.rst
@@ -1,0 +1,34 @@
+interfaces.machine\_parameters package
+========================================================
+
+interfaces.machine\_parameters.ring module
+------------------------------------------------------------
+
+.. automodule:: blond_common.interfaces.machine_parameters.ring
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    
+interfaces.machine\_parameters.ring\_section module
+------------------------------------------------------------
+
+.. automodule:: blond_common.interfaces.machine_parameters.ring_section
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+interfaces.machine\_parameters.rf\_parameters module
+----------------------------------------------------------------------
+
+.. automodule:: blond_common.interfaces.machine_parameters.rf_parameters
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+interfaces.machine\_parameters.rf\_parameters\_options module
+-------------------------------------------------------------------------------
+
+.. automodule:: blond_common.interfaces.machine_parameters.rf_parameters_options
+    :members:
+    :undoc-members:
+    :show-inheritance:
