+++ interfaces/machine_parameters/ring.py
